chapter_004/src/test/java/ru/kbond/collections/list/SimpleStackTest.java
package ru.kbond.collections.list;

import org.junit.Test;
import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.is;

/**
 * Test.
 * @author kbondarenko
 * @since 27.02.2018
 * @version 1
 */
public class SimpleStackTest {
    /**
     * Test.
     */
    @Test
    public void whenAddIntegersThanGetIntegersLIFO() {
        SimpleStack<Integer> stack = new SimpleStack<>();
        stack.push(1);
        stack.push(2);
        stack.push(3);

        assertThat(stack.poll(), is(3));
        assertThat(stack.poll(), is(2));
        assertThat(stack.poll(), is(1));
    }

}